Fix security issues

+++ rosen/modules/render_service_base/src/platform/ohos/rs_render_service_connection_proxy.cpp
@@ -5217,7 +5217,7 @@ ErrCode RSRenderServiceConnectionProxy::SetVirtualScreenStatus(ScreenId id,
         success = false;
         return ERR_INVALID_VALUE;
     }
-    if (!data.WriteUint8(static_cast<uint8_t>(screenStatus))) {
+    if (!data.WriteUint32(static_cast<uint32_t>(screenStatus))) {
         ROSEN_LOGE("SetVirtualScreenStatus: WriteUint8 screenStatus err.");
         success = false;
         return ERR_INVALID_VALUE;
